requirements.txt
ultralytics
streamlit
pillow
opencv-python
torch
